Separate core types from continuation
